Move home api functions to study api file and change study response type
